requirements.txt
requests
openai
moviepy
pillow
imageio-ffmpeg
python-dotenv
